% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing_functions.R
\name{extract_column_from_csc_matrix}
\alias{extract_column_from_csc_matrix}
\title{Extract column from compressed sparse column matrix}
\usage{
extract_column_from_csc_matrix(
  col_nos,
  csc_mat,
  zero_based_idx = TRUE,
  row_subset = NULL
)
}
\arguments{
\item{col_nos}{idxs of the columns to extract}

\item{csc_mat}{a list representing a matrix stored in csc form containing the entries p, i, x, and dim.}

\item{zero_based_idx}{(boolean) do i and p use zero-based indexing}

\item{row_subset}{(optional) subset the extracted columns by row index}
}
\value{
a matrix containing the extracted columns
}
\description{
Extract column from compressed sparse column matrix
}
