% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing_functions.R
\name{extract_row_from_csr_matrix}
\alias{extract_row_from_csr_matrix}
\title{Extract row from compressed sparse row matrix}
\usage{
extract_row_from_csr_matrix(
  row_nos,
  csr_mat,
  zero_based_idx = TRUE,
  col_subset = NULL
)
}
\description{
This function is similar to extract_column_from_csc_matrix; see the documentation of that function.
}
