% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing_functions.R
\name{extract_vector_from_compressed_matrix}
\alias{extract_vector_from_compressed_matrix}
\title{Extract vector from compressed matrix}
\usage{
extract_vector_from_compressed_matrix(
  vector_nos,
  x,
  i,
  p,
  vector_length,
  zero_based_idx = TRUE,
  vector_subset = NULL
)
}
\arguments{
\item{vector_nos}{indexes of the vectors to extract}

\item{x}{sparse-matrix data}

\item{i}{sparse-matrix indexes}

\item{p}{sparse-matrix pointers}

\item{vector_length}{length of the vectors to extract}

\item{zero_based_idx}{(boolean; default TRUE) do i and p use zero-based indexing?}

\item{vector_subset}{(optional) integers indicating a subset of the vectors to return}
}
\value{
the extracted vectors
}
\description{
A function to extract a vector(s) from a matrix stored in compressed column or row format
}
