% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functs.R
\name{get_all_data_for_gene_gRNA}
\alias{get_all_data_for_gene_gRNA}
\title{Get all data for gene-gRNA pair on Gasperini data}
\usage{
get_all_data_for_gene_gRNA(
  gene_id,
  gRNA_id,
 
    sceptre_function_args = "/Users/timbarry/Box/SCEPTRE/sceptre_paper/sceptre_paper/analysis_drivers/analysis_drivers_gasp/sceptre_function_args.R",
  offsite_dir = "/Volumes/tims_new_drive/research/sceptre_files"
)
}
\arguments{
\item{gene_id}{id of the gene}

\item{gRNA_id}{od of the gRNA}

\item{sceptre_function_args}{path to the sceptre_function_args.R file. Defaults to the location on Tim's machine.}

\item{offsite_dir}{path to the offsite dir. Defaults to the location on Tim's machine.}
}
\value{
a list containing expressions, gRNA indicators, gRNA precomp, gene precomp size, gene precomp offsets.
}
\description{
Get all data for gene-gRNA pair on Gasperini data
}
\examples{
gene_id <- "ENSG00000008256"
gRNA_id <- "ACTB_TSS"
get_all_data_for_gene_gRNA(gene_id, gRNA_id)
}
