% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scMAGeCK_functions.R
\name{getsolvedmatrix}
\alias{getsolvedmatrix}
\title{Get solved matrix}
\usage{
getsolvedmatrix(Xm, Ym, lambda = 0.01)
}
\arguments{
\item{Xm}{the X matrix for the regression}

\item{Ym}{the Y matrix}

\item{lambda}{the ridge penalty}
}
\value{

}
\description{
This function was copied and pased from the scMAGeCK package; Yang et al. wrote the function, not Katsevich et al.
}
