% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scMAGeCK_functions.R
\name{getsolvedmatrix_with_permutation_cell_label}
\alias{getsolvedmatrix_with_permutation_cell_label}
\title{getsolvedmatrix_with_permutation_cell_label}
\usage{
getsolvedmatrix_with_permutation_cell_label(
  Xm,
  Ym,
  lambda = 0.01,
  npermutation = 1000
)
}
\arguments{
\item{Xm}{the X matrix for the regression}

\item{Ym}{the Y matrix}

\item{lambda}{the ridge penalty}

\item{npermutation}{number of permutations to run}
}
\value{

}
\description{
This function was copied and pased from the scMAGeCK package; Yang et al. wrote the function, not Katsevich et al.
}
