% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/at_scale_functions.R
\name{map_reduce_at_scale}
\alias{map_reduce_at_scale}
\title{Map reduce at scale}
\usage{
map_reduce_at_scale(
  f,
  pod_id,
  dictionary,
  results_dir,
  cell_gene_expression_matrix,
  ordered_gene_ids,
  gRNA_indicator_matrix_fp,
  covariate_matrix,
  cell_subset = NULL,
  log_dir = NULL
)
}
\arguments{
\item{f}{function we are applying}

\item{pod_id}{pod id}

\item{dictionary}{dictionary of gRNA-gene pairs}

\item{results_dir}{directory in which to store the result}

\item{cell_gene_expression_matrix}{the cell-by-gene expression matrix}

\item{ordered_gene_ids}{the ordered column nmes of the expression matrix}

\item{gRNA_indicator_matrix_fp}{a file-path to the gRNA indicator matrix}

\item{covariate_matrix}{matrix of covariates}

\item{cell_subset}{integer vector indicating cell subset}

\item{log_dir}{location to sink log fles}
}
\description{
Map reduce at scale
}
