% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functs.R
\name{plot_skew_t}
\alias{plot_skew_t}
\title{Plot skew-t}
\usage{
plot_skew_t(resampled_zvalues, original_zvalue, dp)
}
\arguments{
\item{resampled_zvalues}{the set of resampled z-values}

\item{original_zvalue}{the z value of the original negative binomial fit}

\item{dpv}{the skew-t fit MLE}
}
\value{
a ggplot object containing the plot
}
\description{
Plot skew-t
}
