% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functs.R
\name{plot_skew_t_gene_gRNA}
\alias{plot_skew_t_gene_gRNA}
\title{Plot skew-t given gene and gRNA.}
\usage{
plot_skew_t_gene_gRNA(
  gene_id,
  gRNA_id,
  interval = c(-4, 4),
 
    sceptre_function_args = "/Users/timbarry/Box/SCEPTRE/sceptre_paper/sceptre_paper/analysis_drivers/analysis_drivers_gasp/sceptre_function_args.R",
  offsite_dir = "/Volumes/tims_new_drive/research/sceptre_files"
)
}
\arguments{
\item{gene_id}{id of gene}

\item{gRNA_id}{id of gRNA}

\item{sceptre_function_args}{path to the sceptre_function_args.R file. Defaults to the location on Tim's machine.}

\item{offsite_dirpath}{to the offsite dir. Defaults to the location on Tim's machine.}
}
\value{
a ggplot containing the plot
}
\description{
Plot skew-t given gene and gRNA.
}
\examples{
gene_id <- "ENSG00000008256"
gRNA_id <- "ACTB_TSS"
plot_skew_t_gene_gRNA(gene_id, gRNA_id)
}
