% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functs.R
\name{revlog_trans}
\alias{revlog_trans}
\title{Reverse log transformation}
\usage{
revlog_trans(base = exp(1))
}
\arguments{
\item{base}{base of the log (default e)}
}
\description{
Reverse log transformation
}
