% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_based_models.R
\name{run_NB_model}
\alias{run_NB_model}
\title{Run negative binomial model}
\usage{
run_NB_model(expressions, gRNA_indicators, covariate_matrix)
}
\arguments{
\item{expressions}{a vector of expressions}

\item{gRNA_indicators}{a vector of gRNA indicators}

\item{covariate_matrix}{a covariate matrix}
}
\value{
p-value corresponding to the gRNA indicator
}
\description{
Runs a negative binomial model on a vector of expression data, gRNA indicators, and a covariate matrix.
}
\examples{
offsite_dir <- "/Volumes/tims_new_drive/research/sceptre_files"
processed_dir <- paste0(offsite_dir, "/data/xie/processed")
covariate_matrix <- read.fst(paste0(processed_dir, "/covariate_model_matrix.fst"))
ordered_gene_ids <- readRDS(paste0(processed_dir, "/ordered_genes.RDS"))
cell_gene_expression_matrix <- readRDS(paste0(processed_dir, "/exp_mat_metadata.rds")) \%>\% load_fbm
expressions <- cell_gene_expression_matrix[, which(ordered_gene_ids == "ARL15")]
gRNA_indicators <- paste0(processed_dir, "/gRNA_indicator_matrix.fst") \%>\% read.fst() \%>\% pull()
}
