% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_based_models.R
\name{run_nb_model_at_scale}
\alias{run_nb_model_at_scale}
\title{Run NB model at scale}
\usage{
run_nb_model_at_scale(
  pod_id,
  gene_precomp_dir,
  results_dir_negbin,
  cell_gene_expression_matrix,
  ordered_gene_ids,
  gRNA_indicator_matrix_fp,
  covariate_matrix,
  cell_subset,
  log_dir
)
}
\description{
Runs the negative binomial regression model at scale.
}
