% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_based_models.R
\name{run_NB_model_known_size}
\alias{run_NB_model_known_size}
\title{Run negative binomial model (known size)}
\usage{
run_NB_model_known_size(
  expressions,
  gRNA_indicators,
  covariate_matrix,
  gene_size
)
}
\arguments{
\item{expressions}{vector of expressions}

\item{gRNA_indicators}{vector of gRNA indicators}

\item{covariate_matrix}{covariate matrix}

\item{gene_size}{precomputed gene size}
}
\value{
a left-tailed p-value
}
\description{
Runs NB model on gene-gRNA pair where gene has known size. Returns the one-sided p-value.
}
\examples{
offsite_dir <- "/Volumes/tims_new_drive/research/sceptre_files"
processed_dir <- paste0(offsite_dir, "/data/xie/processed")
covariate_matrix <- read.fst(paste0(processed_dir, "/covariate_model_matrix.fst"))
ordered_gene_ids <- readRDS(paste0(processed_dir, "/ordered_gene_ids.RDS"))
cell_gene_expression_matrix <- readRDS(paste0(processed_dir, "/exp_mat_metadata.rds")) \%>\% load_fbm
expressions <- cell_gene_expression_matrix[, which(ordered_gene_ids == "ENSG00000197530.12")]
gRNA_indicators <- paste0(processed_dir, "/gRNA_indicator_matrix.fst") \%>\% read.fst() \%>\% pull()
gene_size <- readRDS(file = paste0(offsite_dir, "/data/xie/precomp/gene/gene_size_unreg_2.rds"))[["ENSG00000197530.12"]]
}
