% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scMAGeCK_functions.R
\name{run_scMaGECK_analysis}
\alias{run_scMaGECK_analysis}
\title{Run scMaGECK analysis}
\usage{
run_scMaGECK_analysis(expression_matrix, gRNA_indic_matrix)
}
\arguments{
\item{expression_matrix}{a matrix of cell-by-gene expressions}

\item{gRNA_indic_matrix}{a matrix of cell-by-gRNA indicators}
}
\value{
p-values for each gRNA-gene pair
}
\description{
Run scMaGECK analysis
}
\examples{
r <- simulate_crispr_screen_data(n_cells = 1000,
grna_mean_prob = 0.2,
mRNA_mean_expression = 40,
covariate_effects_gRNA = rep(4, 10),
covariate_effects_gene = rep(2, 50),
zero_inflation = 0,
neg_binom_size = 2
)
expression_matrix <- r$cell_by_gene_expressions
gRNA_indic_matrix <- r$cell_by_enhancer_perturbation_indicators
run_scMaGECK_analysis(expression_matrix, gRNA_indic_matrix)
}
