% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scMAGeCK_functions.R
\name{run_scmageck_simulation}
\alias{run_scmageck_simulation}
\title{Run scMAGeCK simulation}
\usage{
run_scmageck_simulation(expressions, gRNA_indic_matrix, lib_sizes)
}
\arguments{
\item{expressions}{a vector of expressions}

\item{gRNA_indic_matrix}{the gRNA indicator matrix}

\item{lib_sizes}{the cell-specific library sizes}
}
\value{
a p-value corresponding the gRNA 1
}
\description{
Run scMAGeCK simulation
}
