% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scMAGeCK_functions.R
\name{single_gene_matrix_regression}
\alias{single_gene_matrix_regression}
\title{Get scaled data}
\usage{
single_gene_matrix_regression(
  targetobj,
  ngctrlgene = c("NonTargetingControlGuideForHuman"),
  indmatrix = NULL,
  high_gene_frac = 0.01,
  selected_genes_list = NULL
)
}
\arguments{
\item{targetobj}{a scaled Seurat object}

\item{ngctrlgene}{the name of a gene or genes that serves as a negative control}

\item{indmatrix}{the matrix of gRNA-gene indicators}

\item{high_gene_frac}{remove expressions above this level}

\item{selected_genes_list}{generate Y matrix for these genes}
}
\value{
and X and Y matrix for regression
}
\description{
Obtain the scaled data matrix from a normalized/scaled Seurat object.
}
\details{
This function was copied and pased from the scMAGeCK package; Yang et al. wrote the function, not Katsevich et al.#' Single gene matrix regression

Obtain the X and Y matrices for a regression.

This function was copied and pased from the scMAGeCK package; Yang et al. wrote the function, not Katsevich et al.
}
